+++ Configuration/TSconfig/GridElements/DocumentSectionGridElement.ts
@@ -9,7 +9,7 @@ tx_gridelements {
 
       flexformDS = FILE:EXT:t3v_content/Configuration/FlexForms/GridElements/DocumentSectionGridElement.xml
 
-      icon = EXT:t3v_content/Resources/Public/Icons/GridElements/DocumentSectionGridElement.svg
+      iconIdentifier = t3v_content-document_section_grid_element
 
       # Use a colored frame as in the record form:
       # `0` = no color