+++ ext_tables.php
@@ -20,6 +20,90 @@ call_user_func(function($namespace, $extkey) {
     ['source' => "EXT:{$extkey}/Resources/Public/Icons/ContentElements/SpacerContentElement.svg"]
   );
 
+  // --- Accordion / Accord Grid Element ---
+
+  $iconIdentifier = 'accordion_grid_element';
+
+  $iconRegistry->registerIcon(
+    "{$extkey}-{$iconIdentifier}",
+    \TYPO3\CMS\Core\Imaging\IconProvider\SvgIconProvider::class,
+    ['source' => "EXT:{$extkey}/Resources/Public/Icons/GridElements/AccordionGridElement.svg"]
+  );
+
+  $iconIdentifier = 'accord_grid_element';
+
+  $iconRegistry->registerIcon(
+    "{$extkey}-{$iconIdentifier}",
+    \TYPO3\CMS\Core\Imaging\IconProvider\SvgIconProvider::class,
+    ['source' => "EXT:{$extkey}/Resources/Public/Icons/GridElements/AccordGridElement.svg"]
+  );
+
+  // --- Column Layout Grid Element ---
+
+  $iconIdentifier = '2_columns_layout_grid_element';
+
+  $iconRegistry->registerIcon(
+    "{$extkey}-{$iconIdentifier}",
+    \TYPO3\CMS\Core\Imaging\IconProvider\SvgIconProvider::class,
+    ['source' => "EXT:{$extkey}/Resources/Public/Icons/GridElements/ColumnLayoutGridElement/2Columns.svg"]
+  );
+
+  $iconIdentifier = '3_columns_layout_grid_element';
+
+  $iconRegistry->registerIcon(
+    "{$extkey}-{$iconIdentifier}",
+    \TYPO3\CMS\Core\Imaging\IconProvider\SvgIconProvider::class,
+    ['source' => "EXT:{$extkey}/Resources/Public/Icons/GridElements/ColumnLayoutGridElement/3Columns.svg"]
+  );
+
+  // --- Document Section Grid Element ---
+
+  $iconIdentifier = 'document_section_grid_element';
+
+  $iconRegistry->registerIcon(
+    "{$extkey}-{$iconIdentifier}",
+    \TYPO3\CMS\Core\Imaging\IconProvider\SvgIconProvider::class,
+    ['source' => "EXT:{$extkey}/Resources/Public/Icons/GridElements/DocumentSectionGridElement.svg"]
+  );
+
+  // --- Panel Grid Element ---
+
+  $iconIdentifier = 'panel_grid_element';
+
+  $iconRegistry->registerIcon(
+    "{$extkey}-{$iconIdentifier}",
+    \TYPO3\CMS\Core\Imaging\IconProvider\SvgIconProvider::class,
+    ['source' => "EXT:{$extkey}/Resources/Public/Icons/GridElements/PanelGridElement.svg"]
+  );
+
+  // --- Timeline / Event Grid Element ---
+
+  $iconIdentifier = 'timeline_grid_element';
+
+  $iconRegistry->registerIcon(
+    "{$extkey}-{$iconIdentifier}",
+    \TYPO3\CMS\Core\Imaging\IconProvider\SvgIconProvider::class,
+    ['source' => "EXT:{$extkey}/Resources/Public/Icons/GridElements/TimelineGridElement.svg"]
+  );
+
+  $iconIdentifier = 'event_grid_element';
+
+  $iconRegistry->registerIcon(
+    "{$extkey}-{$iconIdentifier}",
+    \TYPO3\CMS\Core\Imaging\IconProvider\SvgIconProvider::class,
+    ['source' => "EXT:{$extkey}/Resources/Public/Icons/GridElements/EventGridElement.svg"]
+  );
+
+  // --- Viewport Grid Element ---
+
+  $iconIdentifier = 'viewport_grid_element';
+
+  $iconRegistry->registerIcon(
+    "{$extkey}-{$iconIdentifier}",
+    \TYPO3\CMS\Core\Imaging\IconProvider\SvgIconProvider::class,
+    ['source' => "EXT:{$extkey}/Resources/Public/Icons/GridElements/ViewportGridElement.svg"]
+  );
+
   // === TypoScript ===
 
   \TYPO3\CMS\Core\Utility\ExtensionManagementUtility::addStaticFile($extkey, 'Configuration/TypoScript', 'T3v Content');