+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Notable changes will be documented in this file. The project adheres to [Semanti
 Unreleased
 ----------
 
+* Use Icon Registry for Grid Element icons
 * Updated Grid Elements TSconfig
 * Updated `lightboxCssClass` and `lightboxRelAttribute` constant
 * Updated Flexforms (`Columns Layout` Grid Element)